refactor: unify StreamResponses
these were simply forgotten behind, which lead to responses returning
different headers.
